Move to main config
