+++ api/src/main/java/org/commonjava/indy/conf/IndyConfiguration.java
@@ -83,4 +83,6 @@ public interface IndyConfiguration
     boolean isClusterEnabled();
 
     String getCacheKeyspace();
+
+    String getAffectedGroupsExcludeFilter();
 }
