+++ core/src/main/resources/default-main.conf
@@ -1,5 +1,15 @@
+# passthrough.timeout=300
+# nfc.timeout=300
+# nfc.sweep.minutes=30
+# forkjoinpool.common.parallelism=48
+# cluster.enabled=true
 
+# This is a list of http request headers to add to MDC (default: component-id)
+#mdc.headers =
 
+# Ignore affected groups when doing clean-up for the promotion
+#
+#affected.groups.exclude=^build-\d+
 
 [ssl]
 ######################################################################
@@ -21,15 +31,6 @@ remote.ssl.required = false
 remote.nossl.hosts = localhost,127.0.0.1
 
 ######################################################################
-
-# passthrough.timeout=300
-# nfc.timeout=300
-# nfc.sweep.minutes=30
-# forkjoinpool.common.parallelism=48
-# cluster.enabled=true
-
-# This is a list of http request headers to add to MDC (default: component-id)
-#mdc.headers =
 
 # Include addon-specific configurations (or really any configuration) from:
 Include conf.d/*.conf
