+++ deployments/launcher/src/main/etc/indy/main.conf
@@ -1,6 +1,20 @@
+# passthrough.timeout=300
+# nfc.timeout=300
+# nfc.sweep.minutes=30
+# forkjoinpool.common.parallelism=48
+# cluster.enabled=true
 
+# This controls if indy will do downloading when listing a directory content for a remote repo.
+# If it is disabled, indy will only list the content that has been cache in local for remote repo.
+# Default is not enabled.
+# remote.list.download.enabled = false
 
+# This is a list of http request headers to add to MDC (default: component-id)
+#mdc.headers =
 
+# Ignore affected groups when doing clean-up for the promotion
+#
+#affected.groups.exclude=^build-\d+
 
 [ssl]
 ######################################################################
@@ -22,20 +36,6 @@ remote.ssl.required = true
 remote.nossl.hosts = localhost,127.0.0.1
 
 ######################################################################
-
-# passthrough.timeout=300
-# nfc.timeout=300
-# nfc.sweep.minutes=30
-# forkjoinpool.common.parallelism=48
-# cluster.enabled=true
-
-# This controls if indy will do downloading when listing a directory content for a remote repo.
-# If it is disabled, indy will only list the content that has been cache in local for remote repo.
-# Default is not enabled.
-# remote.list.download.enabled = false
-
-# This is a list of http request headers to add to MDC (default: component-id)
-#mdc.headers =
 
 [ui]
